// src/app/models/bank.ts
export interface Bank {
  bankId?: number;
  name?: string;
  code?: string;
  country?: string;
  city?: string;
  address?: string;
  phoneNumber?: string;
  email?: string;
  website?: string;
  notes?: string;
  active?: boolean;
  creationDate?: string;
  createdBy?: string;
  lastModifiedDate?: string;
  lastModifiedBy?: string;
}